requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
vaderSentiment>=3.3.2
nltk>=3.8.0
scikit-learn>=1.3.0
psycopg2-binary>=2.9.0
